* loadave.c         * Include missing header         * NSColorExtensions.m         * TimeMonColors.m           float -> CGFloat conversions, NSInteger conversions         * Percentages.h         * Percentages.m           use NSApplicationTerminateReply
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@2228 d25b4930-3292-e311-b074-0010c6de5da8

+++ ported-apps/Util/TimeMon/NSColorExtensions.m
@@ -6,7 +6,7 @@
 @implementation NSColor (GetColorsFromString)
 + (NSColor *)colorFromStringRepresentation:(NSString *)colorString
 {
-    float r, g, b, a;
+    CGFloat r, g, b, a;
     NSArray *array = [colorString componentsSeparatedByString:@" "];
 
     if (!array) 
@@ -31,7 +31,7 @@
 
 - (NSString *)stringRepresentation
 {
-    float r, g, b, a;
+    CGFloat r, g, b, a;
 
     [[self colorUsingColorSpaceName:NSCalibratedRGBColorSpace] getRed:&r green:&g blue:&b alpha:&a];
     return [NSString stringWithFormat:@"%f %f %f %f",r,g,b,a];
